api/requirements.txt
flask==3.1.1
flask-cors==5.0.1
groq==0.25.0
python-dotenv==1.1.0
